moved metadata function
